Use reference implementation for mnemonic generation.
