+++ README.md
@@ -10,7 +10,12 @@ Create a virtual environment and install the dependencies:
 python3 -m venv ./.venv
 source ./.venv/bin/activate
 pip install -r ./requirements.txt --upgrade
-pip install autopep8
+```
+
+Install development dependencies, as well:
+
+```
+pip install autopep8 pytest
 ```
 
 Above, `requirements.txt` should mirror the **dependencies** section of `pyproject.toml`.
