+++ erdpy_wallet/constants.py
@@ -1,0 +1,2 @@
+BIP39_STRENGTH: int = 256
+BIP39_MNEMONIC_WORD_COUNT = 24
