--- erdpy_wallet/generator.py
@@ -1,9 +1,0 @@
-import nacl.encoding
-import nacl.signing
-
-
-def generate_pair():
-    signing_key = nacl.signing.SigningKey.generate()
-    secret_key = bytes(signing_key)
-    pubkey_bytes = bytes(signing_key.verify_key)
-    return secret_key, pubkey_bytes
