+++ pyproject.toml
@@ -22,6 +22,7 @@ dependencies = [
   "pynacl==1.5.0",
   "pycryptodomex==3.15.0",
   "cryptography==36.0.2",
+  "mnemonic==0.20",
   "erdpy_core @ https://git@github.com/ElrondNetwork/sdk-erdpy-eggs-core.git@init#egg=erdpy_core"
 ]
