+++ requirements.txt
@@ -1,4 +1,5 @@
 pycryptodomex==3.15.0
 cryptography==36.0.2
 pynacl==1.5.0
+mnemonic==0.20
 git+https://git@github.com/ElrondNetwork/sdk-erdpy-eggs-core.git@init#egg=erdpy_core
